# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Entitlements API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/entitlements/broadcasters": {
            "post": {
                "summary": "Register new broadcaster",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\BroadcastersPlansController::registerBroadcaster",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Id of broadcaster which should be registered",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/broadcasters/{broadcasterId}/plans/{planId}/parents": {
            "put": {
                "summary": "Attach plan to parent",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\BroadcastersPlansController::attachPlanToParent",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster which owns plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "planId",
                        "in": "path",
                        "description": "Id of plan which should be attached to parent",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parentPlanId"
                                ],
                                "properties": {
                                    "parentPlanId": {
                                        "description": "Id of desired parent plan",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/broadcasters/{broadcasterId}/plans/{planId}/parents/{parentPlanId}": {
            "delete": {
                "summary": "Detach plan from parent",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\BroadcastersPlansController::detachPlanFromParent",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster which owns plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "planId",
                        "in": "path",
                        "description": "Id of plan which should be detached from parent",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "parentPlanId",
                        "in": "path",
                        "description": "Id of parent plan from which plan should be detached",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/broadcasters/{broadcasterId}/plans/{planId}/categories": {
            "put": {
                "summary": "Assign category to plan",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\BroadcastersPlansController::assignCategoryToPlan",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster which owns plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "planId",
                        "in": "path",
                        "description": "Id of plan to which category should be assigned",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categoryId"
                                ],
                                "properties": {
                                    "categoryId": {
                                        "description": "Id of category which should be assigned to plan",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/broadcasters/{broadcasterId}/plans/{planId}/categories/{categoryId}": {
            "delete": {
                "summary": "Unassign category from plan",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\BroadcastersPlansController::unassignCategoryFromPlan",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster which owns plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "planId",
                        "in": "path",
                        "description": "Id of plan from which category should be unassigned",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id of category which should be unassigned from plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}